Core: extract MapBackedContentStats to a top-level class
Move MapBackedContentStats out of TrackedFileAdapters into its own file,
mirroring the top-level ContentStatsBackedMap on the read side (and its
already top-level test). MapBackedFieldStats stays a private nested helper
since it reads the parent's stat maps directly. Pure relocation with no
behavior change.
